refactor: extract pure text algorithms into text.ts
